# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/auth/change-password": {
            "patch": {
                "description": "Allows the authenticated user to change their password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Change user password",
                "parameters": [
                    {
                        "description": "Old and new password",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/authapp.ChangePasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Password changed successfully",
                        "schema": {
                            "$ref": "#/definitions/app.ChangePasswordSuccessResponseDoc"
                        }
                    },
                    "default": {
                        "description": "Errors",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorDoc"
                        }
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "description": "Authenticate user and return access and refresh tokens",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "User login",
                "parameters": [
                    {
                        "description": "User login payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/authapp.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "schema": {
                            "$ref": "#/definitions/app.LoginSuccessResponseDoc"
                        }
                    },
                    "default": {
                        "description": "Errors",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorDoc"
                        }
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "description": "Invalidates the current user's authentication token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Logout user",
                "responses": {
                    "200": {
                        "description": "Successfully logged out",
                        "schema": {
                            "$ref": "#/definitions/app.LogoutSuccessResponseDoc"
                        }
                    },
                    "default": {
                        "description": "Errors",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorDoc"
                        }
                    }
                }
            }
        },
        "/api/auth/profile": {
            "get": {
                "description": "Retrieve the profile information of the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Get user profile",
                "responses": {
                    "200": {
                        "description": "Profile retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/app.GetProfileSuccessResponseDoc"
                        }
                    },
                    "default": {
                        "description": "Errors",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorDoc"
                        }
                    }
                }
            }
        },
        "/api/auth/refresh-token": {
            "post": {
                "description": "Generate a new access token using a valid refresh token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Refresh access token",
                "parameters": [
                    {
                        "description": "Refresh token payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/authapp.RefreshTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token refreshed successfully",
                        "schema": {
                            "$ref": "#/definitions/app.RefreshTokenSuccessResponseDoc"
                        }
                    },
                    "default": {
                        "description": "Errors",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorDoc"
                        }
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "description": "Create a new user account with email and full name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new userRegisterRequest",
                "parameters": [
                    {
                        "description": "User registration payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/authapp.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User created successfully",
                        "schema": {
                            "$ref": "#/definitions/app.RegisterSuccessResponseDoc"
                        }
                    },
                    "default": {
                        "description": "Errors",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorDoc"
                        }
                    }
                }
            }
        },
...
        "/api/upload/logo": {
            "post": {
                "description": "Upload a logo image to storage and return the public URL",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Upload"
                ],
                "summary": "Upload logo file",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Logo file (png, jpg, jpeg)",
                        "name": "logo",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Upload logo success",
                        "schema": {
                            "$ref": "#/definitions/app.UploadLogoSuccessResponseDoc"
                        }
                    },
                    "default": {
                        "description": "Errors",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorDoc"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "app.ChangePasswordSuccessResponseDoc": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "response_code": {
                    "type": "string"
                }
            }
        },
...
        "app.GetProfileSuccessResponseDoc": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/authapp.GetProfileResponse"
                },
                "message": {
                    "type": "string"
                },
                "response_code": {
                    "type": "string"
                }
            }
        },
...
        "app.LoginSuccessResponseDoc": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/authapp.LoginResponse"
                },
                "message": {
                    "type": "string"
                },
                "response_code": {
                    "type": "string"
                }
            }
        },
        "app.LogoutSuccessResponseDoc": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "response_code": {
                    "type": "string"
                }
            }
        },
        "app.RefreshTokenSuccessResponseDoc": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/authapp.RefreshTokenResponse"
                },
                "message": {
                    "type": "string"
                },
                "response_code": {
                    "type": "string"
                }
            }
        },
        "app.RegisterSuccessResponseDoc": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/authapp.RegisterSuccess"
                },
                "message": {
                    "type": "string"
                },
                "response_code": {
                    "type": "string"
                }
            }
        },
...
        "app.UploadLogoSuccessResponseDoc": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/uploadapp.UploadLogoResponse"
                },
                "message": {
                    "type": "string"
                },
                "response_code": {
                    "type": "string"
                }
            }
        },
        "authapp.ChangePasswordRequest": {
            "type": "object",
            "properties": {
                "confirm_password": {
                    "type": "string"
                },
                "new_password": {
                    "type": "string"
                },
                "old_password": {
                    "type": "string"
                }
            }
        },
        "authapp.GetProfileResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "role": {
                    "type": "string"
                }
            }
        },
        "authapp.LoginRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "authapp.LoginResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "expires_in": {
                    "type": "integer"
                },
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "authapp.RefreshTokenRequest": {
            "type": "object",
            "properties": {
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "authapp.RefreshTokenResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "expires_in": {
                    "type": "integer"
                },
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "authapp.RegisterRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "authapp.RegisterSuccess": {
            "type": "object"
        },
...
        "response.ErrorDoc": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "response_code": {
                    "type": "string"
                }
            }
        },
...
        "uploadapp.UploadLogoResponse": {
            "type": "object",
            "properties": {
                "url": {
                    "type": "string"
                }
            }
        }
    }
}